All primary firmware components have been completed (#33)
* testing without arduino done

* set default parameters

* testing

* a

* test3

* a

* simon is awesome

* a

* a

* ah

* al

* ba

* ah

* ah

* ah

* a

* a

* a

* a

* hi

* hi

* might work

* moved most actions in constructor to begin()

* added moving average to pressure readings

* added sensors

* printing pressure mean

* a

* a

* hi

* ah

* aH

* aH

* aH

* sensor polling frequency bug fix

* flow sensor + blower PID working

* progress'

* reorg

* all main pieces are in, just needs more testing

* can cycle between inhalation and exhalation

* Delete settings.json

+++ firmware/arduino/src/Sensors/MainPressureSensor/MainPressureSensor.cpp
@@ -8,8 +8,9 @@ MainPressureSensor::MainPressureSensor() {
 }
 
 unsigned int MainPressureSensor::read() {
-  int mV = map(analogRead(PRESSURE_PIN), 0, 1023, 0, 5000);
+  // mm H2O
+  float mV = map(analogRead(PRESSURE_PIN), 0, 1023, 0, 5000);
   float mA = mV / PRESSURE_SENSE_RESISTOR;
-  return floor((mA -4) * 25.4); // numbers set by datasheet
+  return floor((mA - 4) * 25.4); // numbers set by datasheet
 }
